The lock is set to 60 FPS. Regardless of the power of the device, the number of frames per second will stay around 60-65

+++ Shadow Adventure/include/windows/game.h
@@ -1,3 +1,3 @@
 #pragma once
 
-int game(int, double); // game
+double game(int, double); // game
